# Conflicts:
#	data-service/src/main/java/org/ihtsdo/buildcloud/service/BuildService.java
#	data-service/src/main/java/org/ihtsdo/buildcloud/service/BuildServiceImpl.java

+++ data-service/src/main/java/org/ihtsdo/buildcloud/dao/BuildDAOImpl.java
@@ -557,7 +557,9 @@ public class BuildDAOImpl implements BuildDAO {
 	public void deleteOutputFiles(Build build) {
 		List<String> outputFiles = listOutputFilePaths(build);
 		for (String outputFile : outputFiles) {
-			buildFileHelper.deleteFile(outputFile);
+			if(buildFileHelper.exists(outputFile)) {
+				buildFileHelper.deleteFile(outputFile);
+			}
 		}
 	}
